requirements.txt
pynput>=1.7.6
twilio>=8.0.0
pystray>=0.19.4
Pillow>=10.0.0
